Fix you say? No I say! Not the issue.
Fixed minecraft not updating (if need be) on a modpack update.
Prevent constant updating of modpacks due to force update.

+++ src/net/ftb/gui/LaunchFrame.java
@@ -569,9 +569,11 @@ public class LaunchFrame extends JFrame {
 	public void runGameUpdater(final LoginResponse response) {
 		final String installPath = Settings.getSettings().getInstallPath();
 		final ModPack modpack = ModPack.getPack(modPacksPane.getSelectedModIndex());
-		File temp = new File(new File(Settings.getSettings().getInstallPath(), modpack.getDir()), "version");
-		if(temp.exists()) {
-			temp.delete();
+		if(Settings.getSettings().getForceUpdate()) {
+			File temp = new File(new File(Settings.getSettings().getInstallPath(), modpack.getDir()), "version");
+			if(temp.exists()) {
+				temp.delete();
+			}
 		}
 		initializeMods();
 		MinecraftVersionDetector mvd = new MinecraftVersionDetector();
